% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_conversions.R
\name{dewpoint_to_rh}
\alias{dewpoint_to_rh}
\title{Relative humidity from air temperature and dew point}
\usage{
dewpoint_to_rh(airt, dewt)
}
\arguments{
\item{airt}{air temperature, degC.}

\item{dewt}{dew point temperature, degC.}
}
\value{
relative humidity, percent, capped at 100.
}
\description{
Relative humidity from air temperature and dew point
}
\examples{
dewpoint_to_rh(20, 13.2)
}
