% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_conversions.R
\name{ds2uv}
\alias{ds2uv}
\title{Wind speed and direction to vector components}
\usage{
ds2uv(d, s)
}
\arguments{
\item{d}{direction the wind blows \emph{from}, degrees.}

\item{s}{wind speed, m/s.}
}
\value{
matrix with columns \code{u} (eastward) and \code{v} (northward), m/s.
}
\description{
Wind speed and direction to vector components
}
\examples{
ds2uv(225, 5)
}
