% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_met.R
\name{met_vars}
\alias{met_vars}
\title{Standard meteorological variables, names and units}
\usage{
met_vars()
}
\value{
data frame with \code{variable}, \code{name} and \code{unit} columns.
}
\description{
Standard meteorological variables, names and units
}
\examples{
met_vars()
}
