% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_conversions.R
\name{mslp_from_station}
\alias{mslp_from_station}
\alias{station_from_mslp}
\title{Convert between station and mean sea level pressure}
\usage{
mslp_from_station(prsttn, elev, airt)

station_from_mslp(prmslp, elev, airt)
}
\arguments{
\item{prsttn}{station pressure, Pa.}

\item{elev}{elevation, m.}

\item{airt}{air temperature, degC.}

\item{prmslp}{mean sea level pressure, Pa.}
}
\value{
the converted pressure, Pa.
}
\description{
Convert between station and mean sea level pressure
}
\examples{
mslp_from_station(97300, elev = 280, airt = 15)
}
