% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_conversions.R
\name{rh_to_dewpoint}
\alias{rh_to_dewpoint}
\title{Dew point temperature from air temperature and relative humidity}
\usage{
rh_to_dewpoint(airt, relh)
}
\arguments{
\item{airt}{air temperature, degC.}

\item{relh}{relative humidity, percent (0-100).}
}
\value{
dew point temperature, degC.
}
\description{
Magnus-Tetens with the Alduchov & Eskridge coefficients (Lawrence 2005);
uncertainty about 0.35 degC.
}
\examples{
rh_to_dewpoint(20, 65)
}
