% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_conversions.R
\name{station_pressure}
\alias{station_pressure}
\title{Station pressure from elevation}
\usage{
station_pressure(airt, relh, elev, mslp = 101325)
}
\arguments{
\item{airt}{air temperature, degC.}

\item{relh}{relative humidity, percent.}

\item{elev}{elevation, m above sea level.}

\item{mslp}{mean sea level pressure, Pa (default standard atmosphere).}
}
\value{
station pressure, \strong{Pa}.
}
\description{
Hypsometric equation using virtual temperature, i.e. accounting for the
moisture content of the air.
}
\examples{
station_pressure(15, 70, elev = 280)
}
