% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_conversions.R
\name{uv2ds}
\alias{uv2ds}
\title{Wind vector components to speed and direction}
\usage{
uv2ds(u, v)
}
\arguments{
\item{u, v}{eastward / northward wind components, m/s.}
}
\value{
matrix with columns \code{dir} (degrees the wind blows \emph{from}) and
\code{speed} (m/s).
}
\description{
Wind vector components to speed and direction
}
\examples{
uv2ds(3, 4)
}
