% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_conversions.R
\name{vapour_pressure}
\alias{vapour_pressure}
\title{Actual vapour pressure}
\usage{
vapour_pressure(airt, relh)
}
\arguments{
\item{airt}{air temperature, degC.}

\item{relh}{relative humidity, percent.}
}
\value{
vapour pressure, hPa.
}
\description{
Eqn. C2 of TVA (1972) as used in the DYRESM manual - the formulation
\code{AEME::expand_met()} uses for \code{MET_prvapr}.
}
\examples{
vapour_pressure(20, 65)
}
